Remove standard distribute lock
