+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/java/org/apache/shardingsphere/mode/manager/cluster/coordinator/lock/InterMutexLock.java
@@ -56,7 +56,7 @@ public final class InterMutexLock implements ShardingSphereGlobalLock {
     
     @Override
     public boolean tryLock(final String lockName, final long timeoutMillis) {
-        return innerTryLock(lockName, timeoutMillis);
+        return innerTryLock(lockName, Math.max(timeoutMillis, TimeoutMilliseconds.MIN_TRY_LOCK));
     }
     
     private boolean innerTryLock(final String lockName, final long timeout) {
