+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/java/org/apache/shardingsphere/mode/manager/cluster/coordinator/lock/global/general/ShardingSphereGeneralDistributeLockManager.java
@@ -44,7 +44,7 @@ import java.util.concurrent.ConcurrentHashMap;
  */
 public final class ShardingSphereGeneralDistributeLockManager implements ShardingSphereDistributeLockManager {
     
-    private final Map<String, ShardingSphereGeneralLock> locks = new ConcurrentHashMap<>();;
+    private final Map<String, ShardingSphereGeneralLock> locks = new ConcurrentHashMap<>();
     
     private final LockNodeService lockNodeService = LockNodeServiceFactory.getInstance().getLockNodeService(LockNodeType.GENERAL);
