+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/java/org/apache/shardingsphere/mode/manager/cluster/coordinator/lock/global/general/node/GeneralLockNodeService.java
@@ -15,24 +15,28 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.database.service;
+package org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.general.node;
 
-import org.junit.Test;
+import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.AbstractGlobalLockNodeService;
+import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.util.LockNodeType;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertThat;
-
-public final class DatabaseStandardLockNodeServiceTest {
+/**
+ * General lock node service.
+ */
+public final class GeneralLockNodeService extends AbstractGlobalLockNodeService {
     
-    private static final DatabaseLockNodeService SERVICE = new DatabaseLockNodeService();
-    
-    @Test
-    public void assertGetSequenceNodePath() {
-        assertThat(SERVICE.getSequenceNodePath(), is("/lock/global/database/sequence"));
+    @Override
+    public String getSequenceNodePath() {
+        return PATH_DELIMITER + LOCK_ROOT + PATH_DELIMITER + LOCK_SCOPE_GLOBAL + PATH_DELIMITER + getLockLevel() + PATH_DELIMITER + "sequence";
     }
     
-    @Test
-    public void assertGetLockLevel() {
-        assertThat(SERVICE.getLockLevel(), is("database"));
+    @Override
+    protected String getLockLevel() {
+        return "general";
+    }
+    
+    @Override
+    public LockNodeType getType() {
+        return LockNodeType.GENERAL;
     }
 }
