+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/java/org/apache/shardingsphere/mode/manager/cluster/coordinator/lock/mutex/MutexLock.java
@@ -40,7 +40,7 @@ public interface MutexLock {
     void unlock();
     
     /**
-     * Is locked
+     * Is locked.
      *
      * @return is locked or not
      */
