+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/resources/META-INF/services/org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.LockNodeService
@@ -15,6 +15,6 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.general.service.GeneralLockNodeService
-org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.database.service.DatabaseLockNodeService
-org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.standard.service.StandardLockNodeService
+org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.general.node.GeneralLockNodeService
+org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.database.node.DatabaseLockNodeService
+org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.mutex.node.MutexLockNodeService
