+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/main/resources/META-INF/services/org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.ShardingSphereDistributeLockManager
@@ -17,4 +17,3 @@
 
 org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.general.ShardingSphereGeneralDistributeLockManager
 org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.database.ShardingSphereDatabaseDistributeLockManager
-org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.standard.ShardingSphereStandardDistributeLockManager
