+++ shardingsphere-mode/shardingsphere-mode-type/shardingsphere-cluster-mode/shardingsphere-cluster-mode-core/src/test/java/org/apache/shardingsphere/mode/manager/cluster/coordinator/lock/global/AbstractGlobalStandardLockNodeServiceTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global;
 
-import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.database.service.DatabaseLockNodeService;
+import org.apache.shardingsphere.mode.manager.cluster.coordinator.lock.global.database.node.DatabaseLockNodeService;
 import org.junit.Test;
 
 import java.util.Optional;
